prediction considers all task input to be at least 1 byte to avoid multiplication/division by 0

+++ hadoop-tools/hadoop-posum/src/main/java/org/apache/hadoop/tools/posum/common/records/dataentity/TaskProfile.java
@@ -15,7 +15,7 @@ public interface TaskProfile extends GeneralDataEntity {
 
     void setOutputBytes(Long outputBytes);
 
-    Integer getDuration();
+    Long getDuration();
 
     TaskType getType();
